% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{perplexica_set_url}
\alias{perplexica_set_url}
\title{Set the Perplexica base URL for the current session}
\usage{
perplexica_set_url(url = "http://localhost:3000")
}
\arguments{
\item{url}{Base URL of the running Perplexica instance.}
}
\value{
The URL, invisibly.
}
\description{
Stores the URL in an R option so it does not need to be passed to every
function call. \code{perplexica_start()} calls this automatically.
}
\examples{
perplexica_set_url("http://localhost:3000")
perplexica_set_url("http://my-server:3000")  # remote instance
}
