% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{perplexica_stop}
\alias{perplexica_stop}
\title{Stop the Perplexica Docker container}
\usage{
perplexica_stop()
}
\value{
Exit status of the \verb{docker stop} command, invisibly.
}
\description{
Stop the Perplexica Docker container
}
\examples{
\dontrun{
perplexica_stop()
}
}
